refactor: centralize Decimal utilities for financial precision
- Create decimal_utils.py with TWO_PLACES, ZERO, to_currency(), and
  FinancialJsonEncoder as single source of truth
- Fix JSON serialization to use string output (preserves precision)
- Replace Decimal(0.0) with ZERO constant to avoid float contamination
- Fix reduce() accumulator in qb.py to use Decimal instead of float
- Convert flexepos.py scraped tips data from float to Decimal
- Update all imports to use decimal_utils module
- Add 28 unit tests for decimal utilities
- Update CLAUDE.md with comprehensive Decimal guidelines

+++ src/flexepos.py
@@ -1,6 +1,7 @@
 import calendar
 import datetime
 import logging
+from decimal import Decimal
 from functools import partial
 from time import sleep
 from typing import Any, Optional, cast
@@ -649,7 +650,7 @@ class Flexepos:
                     rows = tips_table.find_all("tr")
                     rv[store] = [
                         [ele.text.strip() for ele in rows[0].find_all("th")[1:]],
-                        [float(x.text.strip()) for x in rows[1].find_all("td")[1:]],
+                        [Decimal(x.text.strip()) for x in rows[1].find_all("td")[1:]],
                     ]
 
         finally:
